Registration and login is working

+++ mistbank_v2.0/src/com/mistbank/model/MistUser.java
@@ -73,7 +73,7 @@ public class MistUser extends MistLogin {
 		return ssn;
 	}
 
-	public void setSsn(int ssn) {
+	public void setSsn(long ssn) {
 		this.ssn = ssn;
 	}
